dbusdeviation: Do not treat empty XML files as errors
Treat them more like `git diff` does — as deleted, and do the full API
comparison against that, generating a list of the interfaces which have
been added or removed compared to the other file.

This makes comparisons between releases which introduce or remove a
whole interface XML file a lot more natural.