build: Require pre-push hook to be installed for dist to succeed
Add a check to dist-hook to see whether the dbus-deviation pre-push hook
is present and executable in the git repository. If it is, that means it
will be used when the maintainer pushes the new release tag to the
remote, and hence the API signature database will be updated and pushed
as well. If it isn’t, the database would end up getting stale.

This is a bit obnoxious (although it will only be encountered if the
maintainer didn’t fully follow the installation instructions for
dbus-deviation, or if a new maintainer makes a release), but is the only
way I can think of to guarantee that the remote API signature database
will be updated.