dbusapi: Add D-Bus type parser
This is heavily based on a patch by Kaloyan Tenchov <zayfod@gmail.com>,
thanks!

This adds a new Types module, which contains classes for building an AST
to represent a type string; and it adds a TypeParser class, which allows
a type string such as ‘a{sv}’ to be parsed into the AST:
    types.Array
     └─types.DictEntry
        ├─types.String
        └─types.Variant

Some tests are included.

This is a minor API break for ast.Property.type and ast.Argument.type,
which now have a type from the types module rather than a string. The
string representation of this type should be as before though.

https://github.com/pwithnall/dbus-deviation/pull/13